fix: fix bugs in dispense drink service

+++ src/main/java/es/kiwi/drinksdispenser/domain/output/CoinsOutput.java
@@ -4,6 +4,7 @@ import es.kiwi.drinksdispenser.domain.model.Coins;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface CoinsOutput {
@@ -13,7 +14,7 @@ public interface CoinsOutput {
 
     Mono<Void> saveAll(List<Coins> updatedCoins);
 
-    Mono<Double> consultTotalMoneyInMachine(Long machineId);
+    Mono<BigDecimal> consultTotalMoneyInMachine(Long machineId);
 
     Mono<Void> reduceCoins(List<Coins> coins);
 }
